Add files via upload
A freeRTOS based traffic light project. In this project two lights are synchronized together to control the access of two crossed roads. The pedestrian light of that project permits the people to cross the road after the cars are done or the lights are turned to red. For crossing the road it is mandatory to press the pedestrian button. This project will be treated with more functions.